Estruturacao do projeto, pequenas correcoes
